+++ src/components/date-picker/DepthToggle.tsx
@@ -1,11 +1,10 @@
 "use client";
 
 import React, { useState } from "react";
-import SenifitToggleButtonGroup, {
-  ISenifitToggleOption,
-} from "../SenifitToggleButtonGroup";
+import SenifitToggleButtonGroup from "../SenifitToggleButtonGroup";
 import { Box, Typography } from "@mui/material";
 import useMedia from "@/hooks/useMedia";
+import { ISenifitToggleOption } from "@/types/IToggleButton";
 
 export type Depth = "year" | "month" | "day";
